refactor: Remove old code component
