--- design-system/components/code/code-line-number.tsx
@@ -1,4 +1,0 @@
-import { styled } from '@/styled-system/jsx'
-import { codeLineNumber } from './code-line-number.styles'
-
-export const CodeLineNumber = styled('span', codeLineNumber)
