--- design-system/components/code/code-line.tsx
@@ -1,4 +1,0 @@
-import { styled } from '@/styled-system/jsx'
-import { codeLine } from './code-line.styles'
-
-export const CodeLine = styled('div', codeLine)
