--- design-system/components/code/code-token.styles.tsx
@@ -1,9 +1,0 @@
-import { cva } from '@/styled-system/css'
-import { sharedTransitionProperties } from '@/design-system/utils'
-
-export const codeToken = cva({
-  base: {
-    transitionProperty: 'color',
-    ...sharedTransitionProperties
-  }
-})
