--- design-system/components/code/code-token.tsx
@@ -1,4 +1,0 @@
-import { styled } from '@/styled-system/jsx'
-import { codeToken } from './code-token.styles'
-
-export const CodeToken = styled('span', codeToken)
