+++ design-system/components/code/index.tsx
@@ -1,4 +1,3 @@
 export { Code } from './code'
-export { CodeServer } from './code-server'
 export { preToCodeProps } from './code.helpers'
 export { type PreChildrenProps, type PreProps } from './code.types'
